implements new method added in Twitter4J 3.0.5

+++ src/main/scala/twitter4s/Status.scala
@@ -52,7 +52,7 @@ trait Status extends TwitterResponse with EntitySupport {
   
   def contributors: Array[Long]
   
-  def retweetCount: Long
+  def retweetCount: Int
   
   def isRetweetedByMe: Boolean
